src/main/java/Game.java
import java.util.HashMap;

public class Game {
    private HashMap<String, Player> playersMap = new HashMap<>();

    public void register(Player player, String nickname) {
        playersMap.put(nickname, player);
    }

    public int round(String playerName1, String playerName2) {
        Player player1 = null;
        Player player2 = null;

        for (String nickname : playersMap.keySet()) {
            Player player = playersMap.get(nickname);
            if (player.getName().equals(playerName1)) {
                player1 = player;
            }
            if (player.getName().equals(playerName2)) {
                player2 = player;
            }
        }
        if (player1 == null) {
            throw new NotRegisteredException(playerName1);
        }
        if (player2 == null) {
            throw new NotRegisteredException(playerName2);
        }
        if (player1.getStrength() > player2.getStrength()) {
            return 1;
        }
        if (player1.getStrength() < player2.getStrength()) {
            return 2;
        } else {
            return 0;
        }
    }
}

